+++ immutable-arrays/core/src/main/kotlin/com/danrusu/pods4k/immutableArrays/ImmutableArray.kt
@@ -276,6 +276,13 @@ public value class ImmutableArray<out T> @PublishedApi internal constructor(
     }
 
     /**
+     * See [Array.indexOfFirst]
+     */
+    public inline fun indexOfFirst(predicate: (element: T) -> Boolean): Int {
+        return values.indexOfFirst(predicate)
+    }
+
+    /**
      * Creates a pair of immutable arrays, where the first contains elements for which predicate
      * yielded true, and the second contains the other elements.
      */